Set token expiration to 30 minutes

+++ app/Http/Controllers/Api/AuthController.php
@@ -36,7 +36,7 @@ class AuthController extends Controller
         return $this->success([
             'access_token' => $token,
             'token_type' => 'Bearer',
-            'expires_at' => Carbon::now()->addMinutes(config('sanctum.expiration'))->toDateTimeString()
+            'expires_at' => Carbon::now()->addMinutes((int) config('sanctum.expiration'))->toDateTimeString()
         ], 'Login successful', 201);
     }
